+++ X4_ComplexCalculator/Main/WorkArea/SaveDataReaders/SaveDataReader2.cs
@@ -20,7 +20,7 @@ class SaveDataReader2(IMessenger messenger, IWorkArea workArea) : SaveDataReader
     /// ファイル読み込み
     /// </summary>
     /// <returns>成功したか</returns>
-    public override bool Load(IProgress<int> progress)
+    public override bool Load(IProgress<double> progress)
     {
         using var conn = new DBConnection(Path);
